update wl implementation of screen
This commit is rather large, It updates the wayland screen
implementation to try screencopy with dmabuf if available, falling
back to shm. Also provides an ability to get screen brightness per
output. It currently averages the brightness but can be useful if we
allow for per-output screen brightness modifications. In order to
accomplish the above wlr-screencopy-unstable-v1 protocol was updated
and wlr-output-management-unstable-v1 and linux-dmabuf-unstable-v1
were included.